requirements/dev.txt
-r base.txt
ipython==8.14.0
psycopg2-binary==2.9.7
python-environ==0.4.54
